pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "browser-task-copilot"
version = "0.1.0"
description = "Guarded browser task control plane with policy-gated execution and replay artifacts."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Inferensys" }
]
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pyyaml>=6.0.1,<7.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/browser_task_copilot"]
